bumping SDL version to 1.2.15 to solve Ubuntu 12.04 compile issues
From the error report from Rune K. Svendsen:
As far as I can tell (from this:
http://opencast.3480289.n2.nabble.com/3rd-party-tools-libSDL-so-undefined-reference-to-XGetRequest-Ubuntu-12-04-td7485828.html)
the error is caused by Ubuntu 12.04 using libX11 1.4.99.1 which
implements a function called _XGetRequest, which the SDL version 1.2 -
used internally by PeerStreamer - doesn't have. Thus when gcc tries to
link to the newer libSDL.so it fails.